update: many fixes
